refactor(frontend): move KPI token inputs to rewards step
