+++ packages/frontend/src/creation-form/components/generic-data.tsx
@@ -12,12 +12,10 @@ import {
 } from "@carrot-kpi/react";
 import {
     MarkdownInput,
-    NumberInput,
     TagsInput,
     NextStepButton,
     DateTimeInput,
     Typography,
-    type NumberFormatValues,
 } from "@carrot-kpi/ui";
 import {
     unixTimestampToDate,
@@ -27,13 +25,10 @@ import {
 import { NoSpecialCharactersTextInput } from "./no-special-characters-text-input";
 import {
     MAX_KPI_TOKEN_DESCRIPTION_CHARS,
-    MAX_KPI_TOKEN_NAME_CHARS,
-    MAX_KPI_TOKEN_SYMBOL_CHARS,
     MAX_KPI_TOKEN_TAGS_COUNT,
     MAX_KPI_TOKEN_TAG_CHARS,
     MAX_KPI_TOKEN_TITLE_CHARS,
 } from "../constants";
-import { formatUnits, parseUnits } from "viem";
 
 interface GenericDataProps {
     t: NamespacedTranslateFunction;
@@ -54,9 +49,6 @@ export const GenericData = ({
     const [descriptionErrorText, setDescriptionErrorText] = useState("");
     const [tagsErrorText, setTagsErrorText] = useState("");
     const [expirationErrorText, setExpirationErrorText] = useState("");
-    const [tokenNameErrorText, setTokenNameErrorText] = useState("");
-    const [tokenSymbolErrorText, setTokenSymbolErrorText] = useState("");
-    const [tokenSupplyErrorText, setTokenSupplyErrorText] = useState("");
     const [disabled, setDisabled] = useState(true);
     const [minimumDate, setMinimumDate] = useState(new Date());
 
@@ -77,8 +69,6 @@ export const GenericData = ({
         setDisabled(
             !state.title ||
                 !state.description ||
-                !state.tokenName ||
-                !state.tokenSymbol ||
                 !state.title.trim() ||
                 state.title.trim().length > MAX_KPI_TOKEN_TITLE_CHARS ||
                 !stripHtml(state.description).trim() ||
@@ -88,14 +78,7 @@ export const GenericData = ({
                 state.tags.length === 0 ||
                 state.tags.length > MAX_KPI_TOKEN_TAGS_COUNT ||
                 !state.expiration ||
-                isUnixTimestampInThePast(state.expiration) ||
-                !state.tokenName.trim() ||
-                state.tokenName.trim().length > MAX_KPI_TOKEN_NAME_CHARS ||
-                !state.tokenSymbol.trim() ||
-                state.tokenSymbol.trim().length > MAX_KPI_TOKEN_SYMBOL_CHARS ||
-                !state.tokenSupply ||
-                !state.tokenSupply ||
-                parseFloat(state.tokenSupply) === 0,
+                isUnixTimestampInThePast(state.expiration),
         );
     }, [state]);
 
@@ -185,59 +168,6 @@ export const GenericData = ({
         [onStateChange, t],
     );
 
-    const handleTokenNameChange = useCallback(
-        (value: string) => {
-            setTokenNameErrorText(
-                !value
-                    ? t("error.erc20.name.empty")
-                    : value.trim().length > MAX_KPI_TOKEN_NAME_CHARS
-                    ? t("error.erc20.name.tooLong", {
-                          chars: MAX_KPI_TOKEN_NAME_CHARS,
-                      })
-                    : "",
-            );
-            onStateChange((state) => ({
-                ...state,
-                tokenName: value,
-            }));
-        },
-        [onStateChange, t],
-    );
-
-    const handleTokenSymbolChange = useCallback(
-        (value: string) => {
-            setTokenSymbolErrorText(
-                !value
-                    ? t("error.erc20.symbol.empty")
-                    : value.trim().length > MAX_KPI_TOKEN_SYMBOL_CHARS
-                    ? t("error.erc20.symbol.tooLong", {
-                          chars: MAX_KPI_TOKEN_SYMBOL_CHARS,
-                      })
-                    : "",
-            );
-            onStateChange((state) => ({
-                ...state,
-                tokenSymbol: value,
-            }));
-        },
-        [onStateChange, t],
-    );
-
-    const handleTokenSupplyChange = useCallback(
-        (value: NumberFormatValues) => {
-            setTokenSupplyErrorText(
-                !value || !value.value || parseUnits(value.value, 18) === 0n
-                    ? t("error.erc20.supply.zero")
-                    : "",
-            );
-            onStateChange((state) => ({
-                ...state,
-                tokenSupply: parseUnits(value.value, 18).toString(),
-            }));
-        },
-        [onStateChange, t],
-    );
-
     return (
         <div className="flex flex-col gap-6">
             <NoSpecialCharactersTextInput
@@ -302,52 +232,6 @@ export const GenericData = ({
                     inputWrapper: "w-full",
                 }}
             />
-            <div className="flex flex-col md:flex-row w-full gap-4">
-                <NoSpecialCharactersTextInput
-                    label={t("general.label.token.name")}
-                    placeholder={"Example"}
-                    onChange={handleTokenNameChange}
-                    value={state.tokenName}
-                    error={!!tokenNameErrorText}
-                    errorText={tokenNameErrorText}
-                    className={{
-                        root: "w-full",
-                        input: "w-full",
-                        inputWrapper: "w-full",
-                    }}
-                />
-                <NoSpecialCharactersTextInput
-                    label={t("general.label.token.symbol")}
-                    placeholder={"XMPL"}
-                    onChange={handleTokenSymbolChange}
-                    value={state.tokenSymbol}
-                    error={!!tokenSymbolErrorText}
-                    errorText={tokenSymbolErrorText}
-                    className={{
-                        root: "w-full",
-                        input: "w-full",
-                        inputWrapper: "w-full",
-                    }}
-                />
-                <NumberInput
-                    allowNegative={false}
-                    label={t("general.label.token.supply")}
-                    placeholder={"1,000,000"}
-                    onValueChange={handleTokenSupplyChange}
-                    value={
-                        state.tokenSupply
-                            ? formatUnits(BigInt(state.tokenSupply), 18)
-                            : null
-                    }
-                    error={!!tokenSupplyErrorText}
-                    errorText={tokenSupplyErrorText}
-                    className={{
-                        root: "w-full",
-                        input: "w-full",
-                        inputWrapper: "w-full",
-                    }}
-                />
-            </div>
             <NextStepButton onClick={onNext} disabled={disabled}>
                 {t("next")}
             </NextStepButton>
