+++ packages/frontend/src/creation-form/i18n/en.json
@@ -14,9 +14,9 @@
     "general.label.description": "Description",
     "general.label.tags": "Tags",
     "general.label.expiration": "Expiration",
-    "general.label.token.name": "ERC20 token name",
-    "general.label.token.symbol": "ERC20 token symbol",
-    "general.label.token.supply": "ERC20 token supply",
+    "general.label.token.name": "KPI token name",
+    "general.label.token.symbol": "KPI token symbol",
+    "general.label.token.supply": "KPI token supply",
     "general.info.expiration.1": "If the KPI token has not been finalized before this date it will expire and all of the rewards minus any minimum payouts will be sent back to you.",
     "general.info.expiration.2": "Make sure any oracles you configure later will resolve before this date.",
     "general.placeholder.title": "Enter campaign title",
